- added Font Style and Size drops 	- made functional toolbar modular by moving into seperate components 	- still working on functionality of Font Style, Size and lists buttons
